FloatField(default=0)

+++ app_run/views.py
@@ -87,7 +87,7 @@ class PositionViewSet(viewsets.ModelViewSet):
             time_delta = last_position.date_time - previous_position.date_time
             speed = last_distance/time_delta.total_seconds()
             last_position.speed = speed
-            last_position.distance = previous_distance + last_distance/1000
+            last_position.distance = previous_distance + decimal(last_distance/1000)
             last_position.save()
             # print('distance', distance)
             # print('last position', last_position)
